Implement basic wall collision
She at least stops moving at walls but the detection point is a single pixel so it's wonky right now

+++ internal/game/types.go
@@ -10,5 +10,5 @@ const (
 )
 
 type Position struct {
-	x, y int
+	x, y float64
 }
